feat: excel - make new file, make sheet from scratch
